add_hero funciton added. prompt is working
